% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.hobo}
\alias{plot.hobo}
\title{Create an interactive dygraph/plot of processed HOBO logger data}
\usage{
\method{plot}{hobo}(hobo, stn = NULL, flag = 36, save = FALSE,
  out_dir = NULL, verbose = TRUE)
}
\arguments{
\item{hobo}{hobo data.frame created by \code{\link{read_hobo_csv}}}

\item{stn}{character string indicating a relatively short, but descriptive name for the
data logger location. This must be specified. See examples.}

\item{flag}{number of daily measurements *below which* temperature calculations are
flagged to indicate a potentially inadequate number of observations. Default (36) flags
daily values based on 75\% or less of expected observations (i.e., 48 30-min observations
per day). Flagged days are indicated by a red bar along the x-axis.}

\item{save}{logical (default `FALSE`) indicating whether to save the time series plot
to a standalone `html` file named based on `stn` argument. The `html` is saved to
the current working directory unless `out_dir` is specified.}

\item{out_dir}{character string indicating the desired directory to output the `html`
file if `save = TRUE`. Defaults to current working directory.}

\item{verbose}{logical (default `TRUE`) indicating whether to suppress messaging
indicating where output `html` file was saved, if `save = TRUE`#'}
}
\value{
a \code{\link[dygraphs]{dygraph}} interactive time series plot of daily average,
 minimum, and maximum stream temperatures
}
\description{
Create an interactive dygraph/plot of processed HOBO logger data
}
\examples{
\dontrun{
test <- read_hobo_csv(system.file("extdata", "test.csv", package = "r4streams"))
plot(test, "Test deployment")
plot(test, "Another test", save = TRUE, out_dir = "./test_output")
plot(test, "Third test, slightly more descriptive", save = TRUE,
     out_dir = "./test_output")
}
}
